Migrate API calls to FacebookClientWrapper
